LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-setLocalDescription-answer-expected.txt


PASS setLocalDescription() with valid answer should succeed
FAIL setLocalDescription() with type answer and null sdp should use lastAnswer generated from createAnswer promise_test: Unhandled rejection with value: object "OperationError: Expect line: v="
FAIL setLocalDescription() with answer not created by own createAnswer() should reject with InvalidModificationError promise_rejects_dom: function "function () { throw e }" threw object "OperationError: Failed to set local answer sdp: Failed to apply the description for m= section with mid='0': Local fingerprint does not match identity. Expected: sha-256 B0:1B:43:87:8A:C5:01:AD:5C:15:79:5B:A1:94:42:53:14:F4:3D:2A:BA:37:D1:5B:45:9C:2F:F9:3E:9D:13:0B Got: sha-256 FB:6E:CE:7C:14:FE:C1:53:6A:95:EB:0E:9F:71:79:46:F1:67:42:90:97:CD:C1:41:5E:7E:C8:BB:C7:44:30:86" that is not a DOMException InvalidModificationError: property "code" is equal to 0, expected 13
FAIL Calling setLocalDescription(answer) from stable state should reject with InvalidModificationError promise_rejects_dom: function "function () { throw e }" threw object "InvalidStateError: Failed to set local answer sdp: no pending remote description." that is not a DOMException InvalidModificationError: property "code" is equal to 11, expected 13
FAIL Calling setLocalDescription(answer) from have-local-offer state should reject with InvalidModificationError promise_rejects_dom: function "function () { throw e }" threw object "InvalidStateError: Failed to set local answer sdp: no pending remote description." that is not a DOMException InvalidModificationError: property "code" is equal to 11, expected 13
PASS Setting previously generated answer after a call to createOffer should work
FAIL setLocalDescription(answer) should update internal state with a queued task, in the right order assert_equals: pendingRemoteDescription should be updated before the signalingstatechange event expected null but got object "[object RTCSessionDescription]"
